// kservices_ts_api/src/api/domain/entities/sme.entity.ts
export type UpdatenotificationRequest = {
  id : number;
  token: string;
  username: string;
};
// ...
export type getNotificationRequest = {
  id : number;
  agentId: number;
  agentEmail: string;
};

export type markAllAsReadRequest = {
  id : number;
  notificationId: string;
  isRead: string;
};